+++ Source/WebKit2/UIProcess/Launcher/gtk/ProcessLauncherGtk.cpp
@@ -29,10 +29,13 @@
 
 #include "Connection.h"
 #include "ProcessExecutablePath.h"
+#include <WebCore/AuthenticationChallenge.h>
 #include <WebCore/FileSystem.h>
+#include <WebCore/NetworkingContext.h>
 #include <WebCore/ResourceHandle.h>
 #include <WebCore/RunLoop.h>
 #include <errno.h>
+#include <glib.h>
 #include <locale.h>
 #include <wtf/text/CString.h>
 #include <wtf/text/WTFString.h>
@@ -41,6 +44,7 @@
 
 #if OS(LINUX)
 #include <sys/prctl.h>
+#include <sys/socket.h>
 #endif
 
 #ifdef SOCK_SEQPACKET
@@ -57,17 +61,6 @@ static void childSetupFunction(gpointer userData)
 {
     int socket = GPOINTER_TO_INT(userData);
     close(socket);
-
-    // Make child process inherit parent's locale.
-    g_setenv("LC_ALL", setlocale(LC_ALL, 0), TRUE);
-}
-
-static void childFinishedFunction(GPid, gint status, gpointer userData)
-{
-    if (WIFEXITED(status) && !WEXITSTATUS(status))
-        return;
-
-    close(GPOINTER_TO_INT(userData));
 }
 
 void ProcessLauncher::launchProcess()
@@ -81,18 +74,26 @@ void ProcessLauncher::launchProcess()
         return;
     }
 
-    String executablePath = m_launchOptions.processType == WebProcess ?
-                            executablePathOfWebProcess() : executablePathOfPluginProcess();
-    CString binaryPath = fileSystemRepresentation(executablePath);
+    String executablePath, pluginPath;
+    CString realExecutablePath, realPluginPath;
+    if (m_launchOptions.processType == WebProcess)
+        executablePath = executablePathOfWebProcess();
+    else {
+        executablePath = executablePathOfPluginProcess();
+        pluginPath = m_launchOptions.extraInitializationData.get("plugin-path");
+        realPluginPath = fileSystemRepresentation(pluginPath);
+    }
+
+    realExecutablePath = fileSystemRepresentation(executablePath);
     GOwnPtr<gchar> socket(g_strdup_printf("%d", sockets[0]));
-    char* argv[3];
-    argv[0] = const_cast<char*>(binaryPath.data());
+    char* argv[4];
+    argv[0] = const_cast<char*>(realExecutablePath.data());
     argv[1] = socket.get();
-    argv[2] = 0;
+    argv[2] = const_cast<char*>(realPluginPath.data());
+    argv[3] = 0;
 
     GOwnPtr<GError> error;
-    int spawnFlags = G_SPAWN_LEAVE_DESCRIPTORS_OPEN | G_SPAWN_DO_NOT_REAP_CHILD;
-    if (!g_spawn_async(0, argv, 0, static_cast<GSpawnFlags>(spawnFlags), childSetupFunction, GINT_TO_POINTER(sockets[1]), &pid, &error.outPtr())) {
+    if (!g_spawn_async(0, argv, 0, G_SPAWN_LEAVE_DESCRIPTORS_OPEN, childSetupFunction, GINT_TO_POINTER(sockets[1]), &pid, &error.outPtr())) {
         g_printerr("Unable to fork a new WebProcess: %s.\n", error->message);
         ASSERT_NOT_REACHED();
     }
@@ -100,20 +101,22 @@ void ProcessLauncher::launchProcess()
     close(sockets[0]);
     m_processIdentifier = pid;
 
-    // Monitor the child process, it calls waitpid to prevent the child process from becomming a zombie,
-    // and it allows us to close the socket when the child process crashes.
-    g_child_watch_add(m_processIdentifier, childFinishedFunction, GINT_TO_POINTER(sockets[1]));
-
     // We've finished launching the process, message back to the main run loop.
     RunLoop::main()->dispatch(bind(&ProcessLauncher::didFinishLaunchingProcess, this, m_processIdentifier, sockets[1]));
 }
 
 void ProcessLauncher::terminateProcess()
-{   
+{
+    if (m_isLaunching) {
+        invalidate();
+        return;
+    }
+
     if (!m_processIdentifier)
         return;
 
     kill(m_processIdentifier, SIGKILL);
+    m_processIdentifier = 0;
 }
 
 void ProcessLauncher::platformInvalidate()
